requirements.txt
requests~=2.29.0
websockets~=10.4
colorama~=0.4
aiohttp~=3.8.4
brotli~=1.0.9
minestat~=2.6.1
markupsafe~=2.1.2
werkzeug~=2.3.4
click~=8.1.3
blinker~=1.6.2
jinja2~=3.1.2
itsdangerous~=2.1.2
pygithub~=1.55
pillow~=9.5.0
peewee~=3.16.2
bs4~=0.0.1
beautifulsoup4~=4.12.2
rcon~=2.1.1
jieba~=0.42.1
art~=5.9